services/crawler/pyproject.toml
[tool.poetry]
name = "crawler"
version = "0.1.0"
description = ""
authors = ["Ricardo Yaben <rmyl@dtu.dk>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = "~3.10"
PySocks = "^1.7.1"
requests = "^2.26.0"
PyYAML = "^6.0"
beautifulsoup4 = "^4.10.0"
grpcio = "^1.43.0"
grpcio-tools = "^1.43.0"
pytimedinput = "^2.0.1"
lxml = "^4.7.1"
tabulate = "^0.8.9"
tldextract = "^3.2.1"
lib = { path = "../lib"}
